+++ src/rust/src/rdataframe/mod.rs
@@ -169,6 +169,14 @@ impl DataFrame {
         List::from_values(iter)
     }
 
+    pub fn dtype_strings(&self) -> Vec<String> {
+        self.0
+            .get_columns()
+            .iter()
+            .map(|s| format!("{}", s.dtype()))
+            .collect()
+    }
+
     pub fn schema(&self) -> List {
         let mut l = self.dtypes();
         l.set_names(self.0.get_column_names()).unwrap();
@@ -234,6 +242,10 @@ impl DataFrame {
 
         r_result_list(robj_list_res)
     }
+    
+    pub fn frame_equal(&self, other: &DataFrame) -> bool {
+        self.0.frame_equal(&other.0)
+    }
 
     pub fn select_at_idx(&self, idx: i32) -> List {
         let expr_result = || -> Result<Series, String> {
@@ -245,6 +257,10 @@ impl DataFrame {
         r_result_list(expr_result)
     }
 
+    pub fn drop_in_place(&mut self, names: &str) -> Series {
+        Series(self.0.drop_in_place(names).unwrap())
+    }
+    
     pub fn select(&mut self, exprs: &ProtoExprArray) -> list::List {
         let exprs: Vec<pl::Expr> = pra_to_vec(exprs, "select");
         LazyFrame(self.lazy().0.select(exprs)).collect()