+++ NEWS.md
@@ -1,5 +1,8 @@
 # polars (development version)
 
+## What's changed
+- `DataFrame` objects can be subsetted using brackets like standard R data frames: `pl$DataFrame(mtcars)[2:4, c("mpg", "hp")]` (#140 @vincentarelbundock)
+
 # polars v0.5.0
 
 ## BREAKING CHANGE
@@ -9,6 +12,7 @@
 - Several new methods for DataFrame, LazyFrame & GroupBy translated (#103, #105 @vincentarelbundock)
 - Doc fixes (#102, #109  @etiennebacher)
 - Experimental opt-in auto completion (#96 @sorhawell)
+- Base R functions work on DataFrame and LazyFrame objects via S3 methods: as.data.frame, as.matrix, dim, head, length, max, mean, median, min, na.omit, names, sum, tail, unique, ncol, nrow (#107 @vincentarelbundock).
 
 ## New Contributors
 - @etiennebacher made their first contribution in #102